✨ feat: Add student exercise tracking, Lexical editor integration, animations, and Docker deployment
- ✅ Add exercise submission and attempt tracking with scoring system
- 📚 Implement lesson progress tracking (attempts, views, best score)
- 📝 Integrate ReactMarkdown for lesson content rendering
- 🎨 Refactor editor toolbar to use DropdownMenu instead of Select
- ✨ Add animated welcome page with GSAP animations and smooth scrolling
- 🚀 Add Docker production setup with nginx, supervisor, and Reverb
- 🎯 Add course progress calculation and recap functionality
- 🗃️ Add Language seeder with emoji flags for 10 languages
- 🧪 Add comprehensive tests for exercises, progress, and lesson content
- 🐛 Fix popover import from radix-ui
- 💬 Improve conversation UI with dropdown menu
- 🎨 Add RadialProgress component for visual progress indicators
- 🔧 Fix editor formatting inconsistencies (quotes, indentation)

+++ database/migrations/2026_02_02_151032_create_exercise_attempts_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->foreignId('exercise_id')->constrained()->cascadeOnDelete();
-            $table->integer('score');
-            $table->integer('max_score');
+            $table->decimal('score', 5, 2);
+            $table->decimal('max_score', 5, 2);
             $table->boolean('is_passed')->default(false);
             $table->json('answers');
             $table->integer('attempt_number')->default(1);
